# src/swagger/documentation.json
{
  "swagger": "2.0",
  "info": {
    "title": "LibMaS API documentation",
    "description": "Final group project for BYU-I CSE341",
    "version": "0.1a"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/books/": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Return all books from the database",
        "responses": {
          "200": {
            "description": "Books retrieved successfully"
          },
          "500": {
            "description": "Internal server error (databse or node)"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "description": "Inserts a new book to the database",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Json containing book data",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Harry Potter and the Goblet of Fire"
                },
                "authorId": {
                  "type": "string",
                  "example": "4"
                },
                "isbn": {
                  "type": "string",
                  "example": "9780439139601"
                },
                "edition": {
                  "type": "string",
                  "example": "Paperback"
                },
                "pubDate": {
                  "type": "string",
                  "example": "2000-07-08"
                },
                "genre": {
                  "type": "string",
                  "example": "Fantasy"
                },
                "cover": {
                  "type": "string",
                  "example": "https://m.media-amazon.com/images/I/91-LL7OnDCL._AC_UF1000,1000_QL80_.jpg"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New book created successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server error (databse or node)"
          }
        }
      }
    },
...
  }
}